get management ip from assign_network
We get ip from host_interface before.There are two problems.
1、if a interface have two kinds of plane，and have two diff ip of the
two plane。we need to get plane ip from assign_network。
2、interface ip maybe diff from plane ip even only one plane in this
interfcace。for example，when we use a
interface as pxe server and as management.the interface ip is 99.99.x.x
as default,but management is 10.43.x.x.
In a word get plane ip from assign_network is the right way.

Change-Id: Ic99272f6cdfb3ba9d208d08f464aa94303ef1966
Signed-off-by: luyao <lu.yao135@zte.com.cn>